Fix hierarchy snapshotting
