+++ src/snapshot/mod.rs
@@ -4,6 +4,7 @@ use seahash::SeaHasher;
 use std::{collections::VecDeque, marker::PhantomData};
 
 mod checksum;
+mod childof_snapshot;
 mod component_checksum;
 mod component_map;
 mod component_snapshot;
@@ -17,6 +18,7 @@ mod set;
 mod strategy;
 
 pub use checksum::*;
+pub use childof_snapshot::*;
 pub use component_checksum::*;
 pub use component_map::*;
 pub use component_snapshot::*;
@@ -243,152 +245,4 @@ impl<For, As> GgrsComponentSnapshot<For, As> {
 /// Returns a hasher built using the `seahash` library appropriate for creating portable checksums.
 pub fn checksum_hasher() -> SeaHasher {
     SeaHasher::new()
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{prelude::*, schedule_systems::handle_requests};
-    use bevy::prelude::*;
-    use ggrs::*;
-    use serde::{Deserialize, Serialize};
-
-    use crate::PlayerInputs;
-
-    struct TestConfig;
-    impl Config for TestConfig {
-        type Input = Input;
-        type State = u8;
-        type Address = usize;
-    }
-
-    #[derive(Serialize, Deserialize, Clone, Copy, Debug, Default, PartialEq, Eq)]
-    enum Input {
-        #[default]
-        None,
-        SpawnChild,
-        DespawnChildren,
-    }
-
-    #[derive(Component, Clone, Copy)]
-    struct Player;
-
-    fn spawn_child(
-        mut commands: Commands,
-        inputs: Res<PlayerInputs<TestConfig>>,
-        player: Single<Entity, With<Player>>,
-    ) {
-        if inputs[0].0 == Input::SpawnChild {
-            commands.spawn(ChildOf(player.entity())).add_rollback();
-        }
-    }
-
-    fn despawn_children(
-        mut commands: Commands,
-        inputs: Res<PlayerInputs<TestConfig>>,
-        player_children: Single<&Children, With<Player>>,
-    ) {
-        if inputs[0].0 == Input::DespawnChildren {
-            for child in *player_children {
-                commands.entity(*child).despawn();
-            }
-        }
-    }
-
-    fn spawn_player(mut commands: Commands) {
-        commands.spawn(Player).add_rollback();
-    }
-
-    #[test]
-    fn test_hirarchy_preservation() {
-        let mut app = App::new();
-        app.add_plugins(MinimalPlugins);
-        app.add_plugins(GgrsPlugin::<TestConfig>::default());
-        app.add_systems(GgrsSchedule, (spawn_child, despawn_children).chain());
-        app.add_systems(Startup, spawn_player);
-        app.update();
-
-        let cell = GameStateCell::default();
-
-        let get_player = |world: &mut World| {
-            world
-                .query_filtered::<Entity, With<Player>>()
-                .single(world)
-                .unwrap()
-        };
-
-        let get_player_children = |world: &mut World| {
-            let Ok(children) = world
-                .query_filtered::<&Children, With<Player>>()
-                .single(world)
-            else {
-                return vec![];
-            };
-
-            children.into_iter().copied().collect::<Vec<Entity>>()
-        };
-
-        let get_child_parent = |world: &mut World| {
-            world
-                .query::<&ChildOf>()
-                .single(world)
-                .ok()
-                .map(|child_of| child_of.0)
-        };
-
-        let save = |world: &mut World, frame: Frame| {
-            handle_requests(
-                vec![GgrsRequest::<TestConfig>::SaveGameState {
-                    cell: cell.clone(),
-                    frame,
-                }],
-                world,
-            );
-        };
-
-        let advance_frame = |world: &mut World, input: Input| {
-            handle_requests(
-                vec![GgrsRequest::<TestConfig>::AdvanceFrame {
-                    inputs: vec![(input, InputStatus::Predicted)],
-                }],
-                world,
-            );
-        };
-
-        let load = |world: &mut World, frame: Frame| {
-            handle_requests(
-                vec![GgrsRequest::<TestConfig>::LoadGameState {
-                    cell: default(),
-                    frame,
-                }],
-                world,
-            );
-        };
-
-        save(app.world_mut(), 0);
-
-        assert_eq!(get_player_children(app.world_mut()), vec![]);
-
-        // advance to frame 1, spawns a child
-        advance_frame(app.world_mut(), Input::SpawnChild);
-        save(app.world_mut(), 1);
-        let initial_child_enitity = get_player_children(app.world_mut())[0];
-        assert_eq!(get_player_children(app.world_mut()).len(), 1);
-
-        // advance to frame 2, despawns the child
-        advance_frame(app.world_mut(), Input::DespawnChildren);
-        save(app.world_mut(), 2);
-        assert_eq!(get_player_children(app.world_mut()).len(), 0);
-
-        // roll back to frame 1
-        load(app.world_mut(), 1);
-
-        // check that che child was restored
-        assert_eq!(get_player_children(app.world_mut()).len(), 1);
-        let child_entity_after_rollback = get_player_children(app.world_mut())[0];
-        assert_ne!(initial_child_enitity, child_entity_after_rollback);
-        assert_eq!(
-            get_player(app.world_mut()),
-            get_child_parent(app.world_mut()).unwrap()
-        );
-    }
 }
